refactor: extract validators and password strength logic into shared services
