+++ src/app/features/auth/services/password-strength.service.ts
@@ -1,0 +1,40 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class PasswordStrengthService {
+
+  getPasswordStrength(password: string): 'weak' | 'medium' | 'strong' {
+    if (!password) return 'weak';
+
+    let score = 0;
+
+    // Length checks
+    if (password.length >= 8) score++;
+    if (password.length >= 12) score++;
+
+    // Character variety checks
+    if (/[a-z]/.test(password)) score++;
+    if (/[A-Z]/.test(password)) score++;
+    if (/[0-9]/.test(password)) score++;
+    if (/[^A-Za-z0-9]/.test(password)) score++;
+
+    if (score <= 3) return 'weak';
+    if (score <= 5) return 'medium';
+    return 'strong';
+  }
+
+  getPasswordStrengthText(strength: 'weak' | 'medium' | 'strong'): string {
+    switch (strength) {
+      case 'weak':
+        return 'Weak password';
+      case 'medium':
+        return 'Good password';
+      case 'strong':
+        return 'Strong password';
+      default:
+        return '';
+    }
+  }
+}
